Rearrange model app
