+++ model/pyproject.toml
@@ -20,6 +20,7 @@ python-binance = "~1.0.19"
 Flask-JWT-Extended = "4.4.4"
 urllib3 = "1.26.15"
 stratestic = "^1.0.0"
+boto3 = "^1.34.48"
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^7.4.4"
