+++ model/service/helpers/__init__.py
@@ -1,3 +1,3 @@
 from model.service.helpers._helpers import (
-    EXECUTION_APP_ENDPOINTS, APP_NAME, convert_signal_to_text, strategies_defaults
+    EXECUTION_APP_ENDPOINTS, APP_NAME
 )
